upload

+++ backend/database/migrations/2025_03_05_073892_create_resources_categories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger("parent_id");
             $table->foreign('parent_id')->references('id')->on('resources_types')->onDelete('cascade');
             $table->string("label");
-            $table->tinyInteger('sequence')->default(1);
+            $table->integer('sequence')->default(1);
             $table->timestamps();
         });
     }
