+++ backend/database/migrations/2025_03_06_070842_create_risk_levels_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('risk_levels', function (Blueprint $table) {
             $table->id();
             $table->string("label");
-            $table->tinyInteger('sequence')->default(1);
+            $table->integer('sequence')->default(1);
             $table->timestamps();
         });
     }
